add trade resolver, if-else parsing, refactoring

+++ src/main/java/com/made/academy/m1/trade/BondTrade.java
@@ -1,9 +1,7 @@
 package com.made.academy.m1.trade;
 
-import java.math.BigDecimal;
-
 public class BondTrade extends Trade {
-    public BondTrade(BigDecimal price) {
+    public BondTrade(double price) {
         super(price);
     }
 }
