+++ src/main/java/com/made/academy/m1/trade/CommoditySpotTrade.java
@@ -1,9 +1,7 @@
 package com.made.academy.m1.trade;
 
-import java.math.BigDecimal;
-
 public class CommoditySpotTrade extends Trade {
-    public CommoditySpotTrade(BigDecimal price) {
+    public CommoditySpotTrade(double price) {
         super(price);
     }
 }
