+++ src/main/java/com/made/academy/m1/trade/IrSwapTrade.java
@@ -1,9 +1,7 @@
 package com.made.academy.m1.trade;
 
-import java.math.BigDecimal;
-
 public class IrSwapTrade extends Trade {
-    public IrSwapTrade(BigDecimal price) {
+    public IrSwapTrade(double price) {
         super(price);
     }
 }
